file structure for middleware and broke out middleware for login and register
